[Refactoring] Move TrailingClosure to its own file
